# poly eval list: faster `eval_multilinear`